requirements.txt
streamlit==1.44.1
dnspython==2.7.0
motor==3.3.2
pymongo==4.6.2
pillow==11.2.1
python-dotenv==1.1.0
matplotlib==3.10.1
certifi==2025.1.31
requests==2.32.3